refactor(core): move `NodeResizer` & `NodeToolbar` into core pkg
