+++ .changeset/merge-resize-toolbar-into-core.md
@@ -1,0 +1,7 @@
+---
+"@vue-flow/core": minor
+"@vue-flow/node-resizer": major
+"@vue-flow/node-toolbar": major
+---
+
+`NodeResizer`, `NodeResizeControl`, and `NodeToolbar` are now exported directly from `@vue-flow/core`, and the resize-control styles ship with `@vue-flow/core/dist/style.css`. The `@vue-flow/node-resizer` and `@vue-flow/node-toolbar` packages are deprecated; install only `@vue-flow/core` going forward.
