+++ docs/components/DocsRepl.vue
@@ -11,7 +11,6 @@ const { vueFlowVersion } = useVueFlow()
 
 let css = `@import 'https://cdn.jsdelivr.net/npm/@vue-flow/core@${vueFlowVersion}/dist/style.css';
 @import 'https://cdn.jsdelivr.net/npm/@vue-flow/core@${vueFlowVersion}/dist/theme-default.css';
-@import 'https://cdn.jsdelivr.net/npm/@vue-flow/node-resizer@latest/dist/style.css';
 
 html,
 body,
@@ -66,8 +65,6 @@ await store.setFiles(
 store.setImportMap({
   imports: {
     '@vue-flow/core': `${location.origin}/vue-flow-core.mjs`,
-    '@vue-flow/node-resizer': `${location.origin}/vue-flow-node-resizer.mjs`,
-    '@vue-flow/node-toolbar': `${location.origin}/vue-flow-node-toolbar.mjs`,
     ...additionalImports,
   },
 })
