--- docs/examples/node-resizer/style.css
@@ -1,1 +1,0 @@
-@import 'https://cdn.jsdelivr.net/npm/@vue-flow/node-resizer@latest/dist/style.css';
