+++ docs/examples/node-toolbar/ToolbarNode.vue
@@ -1,6 +1,5 @@
 <script setup>
-import { Handle, Position, useVueFlow } from '@vue-flow/core'
-import { NodeToolbar } from '@vue-flow/node-toolbar'
+import { Handle, NodeToolbar, Position, useVueFlow } from '@vue-flow/core'
 
 const props = defineProps(['id', 'data'])
