+++ docs/package.json
@@ -16,8 +16,6 @@
     "@vercel/analytics": "^1.3.2",
     "@vercel/speed-insights": "^1.0.14",
     "@vue-flow/core": "workspace:*",
-    "@vue-flow/node-resizer": "workspace:*",
-    "@vue-flow/node-toolbar": "workspace:*",
     "@vue/repl": "3.4.0",
     "blobity": "^0.2.3",
     "vue": "^3.5.12",
