+++ examples/vite/package.json
@@ -9,8 +9,6 @@
   "dependencies": {
     "@dagrejs/dagre": "^1.1.4",
     "@vue-flow/core": "workspace:*",
-    "@vue-flow/node-resizer": "workspace:*",
-    "@vue-flow/node-toolbar": "workspace:*",
     "html-to-image": "^1.11.11",
     "pinia": "^2.2.4"
   },
