+++ packages/core/src/components/NodeResizer/ResizeControlLine.vue
