+++ packages/core/src/components/NodeResizer/index.ts
@@ -1,5 +1,3 @@
-import './style.css'
-
 export { default as NodeResizer } from './NodeResizer.vue'
 export { default as NodeResizeControl } from './ResizeControl.vue'
