+++ packages/core/src/components/NodeResizer/types.ts
