+++ packages/core/src/components/NodeToolbar/index.ts
@@ -1,2 +1,3 @@
-export * from './types'
 export { default as NodeToolbar } from './NodeToolbar.vue'
+
+export * from './types'
