--- packages/node-resizer/package.json
@@ -1,79 +1,0 @@
-{
-  "name": "@vue-flow/node-resizer",
-  "version": "1.5.0",
-  "private": false,
-  "license": "MIT",
-  "author": "Burak Cakmakoglu<78412429+bcakmakoglu@users.noreply.github.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/bcakmakoglu/vue-flow/packages/plugins/node-resizer"
-  },
-  "homepage": "https://github.com/bcakmakoglu/vue-flow#readme",
-  "bugs": {
-    "url": "https://github.com/bcakmakoglu/vue-flow/issues"
-  },
-  "keywords": [
-    "vue",
-    "flow",
-    "diagram",
-    "editor",
-    "graph",
-    "node",
-    "link",
-    "port",
-    "slot",
-    "vue3",
-    "composition-api",
-    "vue-flow",
-    "vueflow",
-    "typescript"
-  ],
-  "main": "./dist/vue-flow-node-resizer.js",
-  "module": "./dist/vue-flow-node-resizer.mjs",
-  "types": "./dist/index.d.ts",
-  "unpkg": "./dist/vue-flow-node-resizer.iife.js",
-  "jsdelivr": "./dist/vue-flow-node-resizer.iife.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/vue-flow-node-resizer.mjs",
-      "require": "./dist/vue-flow-node-resizer.js"
-    },
-    "./dist/style.css": "./dist/style.css"
-  },
-  "files": [
-    "dist",
-    "*.d.ts"
-  ],
-  "sideEffects": [
-    "*.css"
-  ],
-  "scripts": {
-    "dev": "pnpm types:watch & pnpm build:watch",
-    "build": "vite build",
-    "build:watch": "vite build --watch",
-    "types": "vue-tsc --declaration --emitDeclarationOnly && pnpm lint:dist",
-    "types:watch": "vue-tsc --declaration --emitDeclarationOnly --watch",
-    "lint": "eslint --ext .js,.ts,.vue ./",
-    "lint:dist": "eslint --ext \".ts,.tsx\" -c .eslintrc.js --fix --ignore-pattern !**/*  ./dist",
-    "test": "exit 0"
-  },
-  "peerDependencies": {
-    "@vue-flow/core": "^1.23.0",
-    "vue": "^3.3.0"
-  },
-  "dependencies": {
-    "@xyflow/system": "^0.0.73"
-  },
-  "devDependencies": {
-    "@tooling/eslint-config": "workspace:*",
-    "@tooling/tsconfig": "workspace:*",
-    "@tooling/vite-config": "workspace:*",
-    "@vue-flow/core": "workspace:*",
-    "vue-tsc": "^1.8.16"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
-}
