--- packages/node-toolbar/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: ['@tooling/eslint-config'],
-}
