--- packages/node-toolbar/CHANGELOG.md
@@ -1,61 +1,0 @@
-# @vue-flow/node-toolbar
-
-## 1.1.1
-
-### Patch Changes
-
-- [#1785](https://github.com/bcakmakoglu/vue-flow/pull/1785) [`0179b98`](https://github.com/bcakmakoglu/vue-flow/commit/0179b982cc7a74d4aa99b4a5c1ba815e0cd4db28) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Set default value for `isVisible` to `undefined`
-
-## 1.1.0
-
-### Minor Changes
-
-- [#953](https://github.com/bcakmakoglu/vue-flow/pull/953) [`bda15cfa`](https://github.com/bcakmakoglu/vue-flow/commit/bda15cfa790ae479e95827b940c0043448f7f89c) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Add `align` prop to node toolbar
-
-## 1.0.7
-
-### Patch Changes
-
-- [#865](https://github.com/bcakmakoglu/vue-flow/pull/865) [`9ce7bdc4`](https://github.com/bcakmakoglu/vue-flow/commit/9ce7bdc4908dda4dea299e5f469b252ac20a12ab) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Add `exports` field to package.json
-
-## 1.0.6
-
-### Patch Changes
-
-- [#716](https://github.com/bcakmakoglu/vue-flow/pull/716) [`1685827d`](https://github.com/bcakmakoglu/vue-flow/commit/1685827d0ea1dc9864f95a1b3a54fbc43a296e5d) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Fix umd pkg names and use the correct vue flow core umd pkg name in plugins
-
-## 1.0.5
-
-### Patch Changes
-
-- [#682](https://github.com/bcakmakoglu/vue-flow/pull/682) [`b08cb4d4`](https://github.com/bcakmakoglu/vue-flow/commit/b08cb4d45904c229d9ecda5e3cb477cbb7a6acaf) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Add compat mode to components to avoid breaking when used with @vue/compat
-
-## 1.0.4
-
-### Patch Changes
-
-- [#616](https://github.com/bcakmakoglu/vue-flow/pull/616) [`b16e3564`](https://github.com/bcakmakoglu/vue-flow/commit/b16e3564708c5429ad594156341fa3e95f84d3b2) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Upgrade to vite 4 & update deps
-
-## 1.0.3
-
-### Patch Changes
-
-- [`c9d37fc`](https://github.com/bcakmakoglu/vue-flow/commit/c9d37fcad85e6f7643d3905d34a2d2c6566b3977) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Correct pkg main field
-
-## 1.0.2
-
-### Patch Changes
-
-- [#569](https://github.com/bcakmakoglu/vue-flow/pull/569) [`c0ee1cd`](https://github.com/bcakmakoglu/vue-flow/commit/c0ee1cdfdc70ea61766b02de8f8acb4f0cad8bac) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Wait for viewport ref to be available before teleporting to it
-
-## 1.0.1
-
-### Patch Changes
-
-- [#553](https://github.com/bcakmakoglu/vue-flow/pull/553) [`d4e0b7e`](https://github.com/bcakmakoglu/vue-flow/commit/d4e0b7eb1ed1cadab1ec778f2e71cf57ba8cd23e) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Use renamed `ViewportTransform` type
-
-## 1.0.0
-
-### Major Changes
-
-- [#476](https://github.com/bcakmakoglu/vue-flow/pull/476) [`136cc19f`](https://github.com/bcakmakoglu/vue-flow/commit/136cc19fe29a16bc139655f8ce491fe94b1182e8) Thanks [@bcakmakoglu](https://github.com/bcakmakoglu)! - Add node toolbar package
