--- packages/node-toolbar/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@vue-flow/node-toolbar",
-  "version": "1.1.1",
-  "private": false,
-  "license": "MIT",
-  "author": "Burak Cakmakoglu<78412429+bcakmakoglu@users.noreply.github.com>",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/bcakmakoglu/vue-flow/packages/plugins/node-toolbar"
-  },
-  "homepage": "https://github.com/bcakmakoglu/vue-flow#readme",
-  "bugs": {
-    "url": "https://github.com/bcakmakoglu/vue-flow/issues"
-  },
-  "main": "./dist/vue-flow-node-toolbar.js",
-  "module": "./dist/vue-flow-node-toolbar.mjs",
-  "types": "./dist/index.d.ts",
-  "unpkg": "./dist/vue-flow-node-toolbar.iife.js",
-  "jsdelivr": "./dist/vue-flow-node-toolbar.iife.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/vue-flow-node-toolbar.mjs",
-      "require": "./dist/vue-flow-node-toolbar.js"
-    }
-  },
-  "files": [
-    "dist",
-    "*.d.ts"
-  ],
-  "sideEffects": false,
-  "scripts": {
-    "dev": "pnpm types:watch & pnpm build:watch",
-    "build": "vite build",
-    "build:watch": "vite build --watch",
-    "types": "vue-tsc --declaration --emitDeclarationOnly && pnpm lint:dist",
-    "types:watch": "vue-tsc --declaration --emitDeclarationOnly --watch",
-    "lint": "eslint --ext .js,.ts,.vue ./",
-    "lint:dist": "eslint --ext \".ts,.tsx\" -c .eslintrc.js --fix --ignore-pattern !**/*  ./dist",
-    "test": "exit 0"
-  },
-  "peerDependencies": {
-    "@vue-flow/core": "^1.23.0",
-    "vue": "^3.3.0"
-  },
-  "devDependencies": {
-    "@tooling/eslint-config": "workspace:*",
-    "@tooling/tsconfig": "workspace:*",
-    "@tooling/vite-config": "workspace:*",
-    "@vue-flow/core": "workspace:*",
-    "@xyflow/system": "^0.0.73",
-    "vue-tsc": "^1.8.16"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
-}
