--- packages/node-toolbar/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@tooling/tsconfig/base",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "declarationDir": "./dist",
-    "types": [
-      "vite/client",
-      "vue/macros"
-    ]
-  },
-  "include": [
-    "./src"
-  ]
-}
